requirements.txt
aura-sr
fastapi
uvicorn
Pillow